fixed tests and renamed folder names to solve github issue
