+++ index.js
@@ -1,5 +1,5 @@
 import { AppRegistry } from 'react-native';
-import App from './src/app';
+import App from './src/new-app';
 import { name as appName } from './app.json';
 
 AppRegistry.registerComponent(appName, () => App);
