--- src/components/brand/styles.ts
@@ -1,10 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({
-  brandContainer: {
-    display: 'flex',
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-});
