+++ src/components/index.ts
@@ -1,3 +1,3 @@
-export { default as Brand } from './brand/brand';
+export { default as Brand } from './my-brand/my-brand';
 export { default as CustomButton } from './custom-button/custom-button';
 export { default as ErrorFallback } from './error-fallback/error-fallback';
