+++ src/components/my-brand/my-brand.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react-native';
-import Brand from './brand';
+import Brand from './my-brand';
 import { CatContextProvider } from 'boilerplate-react-native/src/contexts';
 
-test('render correctly', () => {
+test('App renders correctly', () => {
   const component = (
     <CatContextProvider>
       <Brand />
@@ -13,9 +13,8 @@ test('render correctly', () => {
   render(component);
 
   const wrapper = screen.getByTestId('brand-img-wrapper');
-  const img = screen.getByTestId('brand-img');
 
-  expect(wrapper.props.style.height).toBe(200);
-  expect(wrapper.props.style.width).toBe(200);
-  expect(img.props.resizeMode).toBe('contain');
+  expect(wrapper.props.style).toHaveLength(2);
+  expect(wrapper.props.style[1].height).toBe(400);
+  expect(wrapper.props.style[1].width).toBe(400);
 });
